refactor: reorganize scheduler functionality into separate modules for better maintainability
